# api-gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for Smart City Management System resources",
        "title": "Smart City Management System API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/planning/demographics/{areaId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get demographics details for a specific area by ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "city_planning"
                ],
                "summary": "Get demographics details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Area ID",
                        "name": "areaId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/city_planning.DemographicDataList"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/planning/infrastructure/{infrastructureId}/status": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get status of a specific infrastructure by ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "city_planning"
                ],
                "summary": "Get infrastructure status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Infrastructure ID",
                        "name": "infrastructureId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/city_planning.InfrastuctureAssetList"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/planning/land-use/{parcelId}/change-request": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a land-use change request for a specific parcel.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "city_planning"
                ],
                "summary": "Create a land-use change request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Parcel ID",
                        "name": "parcelId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/city_planning.Void"
                        }
                    },
                    "400": {
                        "description": "error\": \"Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/planning/permits/{permitTypeId}/requirements": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get requirements for a specific permit type by ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "city_planning"
                ],
                "summary": "Get permit requirements",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Permit Type ID",
                        "name": "permitTypeId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/city_planning.BuildingPermitFilter"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/planning/proposals": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new planning proposal with the provided details.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "city_planning"
                ],
                "summary": "Create a new planning proposal",
                "parameters": [
                    {
                        "description": "Planning Proposal Data",
                        "name": "proposal",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/city_planning.PlanningProposalCreate"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/city_planning.Void"
                        }
                    },
                    "400": {
                        "description": "error\": \"Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/planning/proposals/{proposalId}/feedback": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Submit feedback for a specific planning proposal.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "city_planning"
                ],
                "summary": "Submit feedback for a planning proposal",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Proposal ID",
                        "name": "proposalId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Proposal Feedback Data",
                        "name": "feedback",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/city_planning.ProposalFeedbackCreate"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/city_planning.Void"
                        }
                    },
                    "400": {
                        "description": "error\": \"Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/planning/zones/{zoneId}/current": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get current zone details by zone ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "city_planning"
                ],
                "summary": "Get current zone details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Zone ID",
                        "name": "zoneId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/city_planning.CityZoneList"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "city_planning.BuildingPermitFilter": {
            "type": "object",
            "properties": {
                "ApplicationName": {
                    "type": "string"
                },
                "PermitId": {
                    "type": "string"
                },
                "PermitType": {
                    "type": "string"
                },
                "Status": {
                    "type": "string"
                },
                "ZoneId": {
                    "type": "string"
                }
            }
        },
        "city_planning.CityZone": {
            "type": "object",
            "properties": {
                "Area": {
                    "$ref": "#/definitions/city_planning.Polygon"
                },
                "CreatedAt": {
                    "type": "string"
                },
                "CurrentUsage": {
                    "type": "string"
                },
                "DeletedAt": {
                    "type": "string"
                },
                "UpdatedAt": {
                    "type": "string"
                },
                "ZoneId": {
                    "type": "string"
                },
                "ZoneName": {
                    "type": "string"
                }
            }
        },
        "city_planning.CityZoneList": {
            "type": "object",
            "properties": {
                "CityZones": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/city_planning.CityZone"
                    }
                }
            }
        },
        "city_planning.DemographicData": {
            "type": "object",
            "properties": {
                "DataId": {
                    "type": "string"
                },
                "DataYear": {
                    "type": "string"
                },
                "DeletedAt": {
                    "type": "string"
                },
                "MedianAge": {
                    "type": "integer"
                },
                "MedianIncome": {
                    "type": "number"
                },
                "Population": {
                    "type": "integer"
                },
                "UpdatedAt": {
                    "type": "string"
                },
                "ZoneId": {
                    "type": "string"
                }
            }
        },
        "city_planning.DemographicDataList": {
            "type": "object",
            "properties": {
                "DemographicDatas": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/city_planning.DemographicData"
                    }
                }
            }
        },
        "city_planning.InfrastuctureAsset": {
            "type": "object",
            "properties": {
                "AssetId": {
                    "type": "string"
                },
                "AssetType": {
                    "type": "string"
                },
                "CurrentStatus": {
                    "type": "string"
                },
                "DeletedAt": {
                    "type": "string"
                },
                "InstallationDate": {
                    "type": "string"
                },
                "Location": {
                    "$ref": "#/definitions/city_planning.Point"
                },
                "UpdatedAt": {
                    "type": "string"
                }
            }
        },
        "city_planning.InfrastuctureAssetList": {
            "type": "object",
            "properties": {
                "Assets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/city_planning.InfrastuctureAsset"
                    }
                }
            }
        },
        "city_planning.PlanningProposalCreate": {
            "type": "object",
            "properties": {
                "Description": {
                    "type": "string"
                },
                "ProposalId": {
                    "type": "string"
                },
                "ProposalType": {
                    "type": "string"
                },
                "Status": {
                    "type": "string"
                },
                "ZoneId": {
                    "type": "string"
                }
            }
        },
        "city_planning.Point": {
            "type": "object",
            "properties": {
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                }
            }
        },
        "city_planning.Polygon": {
            "type": "object",
            "properties": {
                "vertices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/city_planning.Point"
                    }
                }
            }
        },
        "city_planning.ProposalFeedbackCreate": {
            "type": "object",
            "properties": {
                "CitizenId": {
                    "type": "string"
                },
                "FeedbackId": {
                    "type": "string"
                },
                "FeedbackText": {
                    "type": "string"
                },
                "ProposalId": {
                    "type": "string"
                }
            }
        },
        "city_planning.Void": {
            "type": "object"
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}